+++ src/Models/Traits/ExtractData.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CTApi\Models\Traits;
+
+trait ExtractData
+{
+    public function extractData(): array
+    {
+        return get_object_vars($this);
+    }
+
+    public abstract function getModifiableAttributes(): array;
+}