+++ src/Requests/AbstractRequestBuilder.php
@@ -5,6 +5,7 @@ namespace CTApi\Requests;
 
 
 use CTApi\CTClient;
+use CTApi\Exceptions\CTModelException;
 use CTApi\Exceptions\CTRequestException;
 use CTApi\Models\AbstractModel;
 use CTApi\Requests\Traits\OrderByCondition;
@@ -61,6 +62,52 @@ abstract class AbstractRequestBuilder
         $this->orderRawData($data);
 
         return $this->getModelClass()::createModelsFromArray($data);
+    }
+
+    /**
+     * Update object's data in ChurchTools by the given object ID.
+     * @param string $objectId
+     * @param array $data Key-Value pair with attributes
+     */
+    protected function updateData(string $objectId, array $data): void
+    {
+        $url = $this->getApiEndpoint() . '/' . $objectId;
+
+        $client = CTClient::getClient();
+        $response = $client->patch($url, ['json' => $data]);
+
+        if ($response->getStatusCode() !== 200) {
+            throw CTRequestException::ofErrorResponse($response);
+        }
+    }
+
+    /**
+     * Send Update-Request for given Model. Only update Attributes that are given with the updateAttributes-Parameter.
+     *
+     * @param $model Model
+     * @param string $modelId Id of Model
+     * @param array $updateAttributes Pass the attributes that should be updated as array. If nothing or
+     *        an empty array is passed, all data of the person will be sent to the API.
+     *        If an array is passed that looks like this:
+     *        <code>['firstName', 'lastName', 'nickname']</code>
+     *        only those attributes will be sent to the API.
+     */
+    protected function updateDataForModel($model, string $modelId, array $updateAttributes): void
+    {
+        if (empty($updateAttributes)) {
+            $updateAttributes = $model->getModifiableAttributes();
+        } else {
+            $diff = array_diff($updateAttributes, $model->getModifiableAttributes());
+
+            if ($diff) {
+                throw new CTModelException('The attributes ' . implode(', ', $diff) . ' of Model ' . get_class($model) . ' are not modifiable.');
+            }
+        }
+
+        $allData = $model->extractData();
+        $updateAttributes = array_intersect_key($allData, array_flip($updateAttributes));
+
+        $this->updateData($modelId, $updateAttributes);
     }
 
     abstract protected function getApiEndpoint(): string;